coding_devotionals_fa23/Slytherin/requirements.txt
streamlit-folium
folium
pandas
numpy
plotly-express